requirements.txt
matplotlib
numpy
pyserial
Pillow
adafruit-circuitpython-bno055
adafruit-circuitpython-pca9685
adafruit-circuitpython-servokit
adafruit-blinka
bleak
